+++ build_modules/lib/src/modules.dart
@@ -113,35 +113,40 @@ class Module extends Object with _$ModuleSerializerMixin {
   /// Generated factory constructor.
   factory Module.fromJson(Map<String, dynamic> json) => _$ModuleFromJson(json);
 
-  /// Computes the [primarySource]s of all [Module]s that are transitively
-  /// depended on by this module.
+  /// Returns all [Module]s in the transitive dependencies of this module in a
+  /// post-order traversal.
   ///
   /// Throws a [MissingModulesException] if there are any missing modules. This
   /// typically means that somebody is trying to import a non-existing file.
   Future<List<Module>> computeTransitiveDependencies(AssetReader reader) async {
-    var transitiveDeps = <AssetId, Module>{};
-    var modulesToCrawl = directDependencies.toSet();
+    var modulesByPrimarySource = <AssetId, Module>{};
+    var postOrderModules = <Module>[];
+    var crawlStack = <AssetId>[]..addAll(directDependencies);
     var missingModuleSources = new Set<AssetId>();
-    while (modulesToCrawl.isNotEmpty) {
-      var next = modulesToCrawl.last;
-      modulesToCrawl.remove(next);
-      if (transitiveDeps.containsKey(next)) continue;
+    while (crawlStack.isNotEmpty) {
+      var next = crawlStack.last;
+      if (modulesByPrimarySource.containsKey(next)) {
+        postOrderModules.add(modulesByPrimarySource[next]);
+        crawlStack.removeLast();
+        continue;
+      }
       var nextModuleId = next.changeExtension(moduleExtension);
       if (!await reader.canRead(nextModuleId)) {
         missingModuleSources.add(next);
+        crawlStack.removeLast();
         continue;
       }
       var module = new Module.fromJson(
           json.decode(await reader.readAsString(nextModuleId))
               as Map<String, dynamic>);
-      transitiveDeps[next] = module;
-      modulesToCrawl.addAll(module.directDependencies);
+      modulesByPrimarySource[next] = module;
+      crawlStack.addAll(module.directDependencies);
     }
     if (missingModuleSources.isNotEmpty) {
       throw await MissingModulesException.create(this, missingModuleSources,
-          transitiveDeps.values.toList()..add(this), reader);
+          modulesByPrimarySource.values.toList()..add(this), reader);
     }
-    return transitiveDeps.values.toList();
+    return postOrderModules;
   }
 }